% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nohrsc-package.R
\docType{package}
\name{nohrsc-package}
\alias{nohrsc}
\alias{nohrsc-package}
\title{nohrsc: Access NOAA National Operational Hydrologic Remote Sensing Center's Snow Analysis Data Archive}
\description{
Simplifies access to the NOHRSC data archive. It takes common name 
    arguments and dates to construct the proper URLs and access the correctly 
    formatted file names. Multiple days and products can be downloaded and 
    extracted in a single call. The extracted rasters will be converted to .bil 
    format and an appropriate .hdr and .prj will be created for each raster.
}
\author{
\strong{Maintainer}: Joe Shannon \email{jpshanno@mtu.edu} (\href{https://orcid.org/0000-0002-0052-9962}{ORCID})

}
\keyword{internal}
