% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsa_get_snodas.R
\name{nsa_get_snodas}
\alias{nsa_get_snodas}
\title{Download and extract national snow analysis data}
\usage{
nsa_get_snodas(
  product,
  start.date,
  end.date = NULL,
  region = "North America",
  path = "",
  delete.archive = FALSE,
  force.download = FALSE
)
}
\arguments{
\item{product}{A character vector of one or more products available via
NOHRSC Acceptable values are "liquid precipitation", "snow precipitation",
"snow water equivalent", "snow depth", "snow pack average temperature",
"blowing snow sublimation", "snow melt", and "snow pack sublimation"}

\item{start.date}{A start date supplied as a character string ("YYYY-MM-DD")
or a datetime object}

\item{end.date}{NULL if only one date is wanted, otherwise the end date of a
time span formatted as start.date}

\item{region}{Either "CONUS" or "North America" indicating if the results
should be contained to the continental US or not}

\item{path}{The output path to save the raster data}

\item{delete.archive}{A logical indicating if the downloaded archives be
deleted after the rasters have been extracted. Defaults to FALSE}

\item{force.download}{A logical indicating if archive should be redownloaded
if the file exists in \code{path}.}
}
\value{
A character vector of the extracted filenames
}
\description{
Download and extract national snow analysis data
}
\examples{
\dontrun{
raster_files <-
  nsa_get_snodas(c("snow water equivalent", "snow melt"),
                 start.date = "2010-01-06",
                 end.date = "2010-01-08",
                 path = "snodas")
}
}
\references{
\url{https://nsidc.org/data/g02158}
}
